refactor: services to similar structure
