+++ packages/ui/src/components/va-config/VaConfig.demo.vue
@@ -75,7 +75,7 @@
 
 <script>
 import { computed } from 'vue'
-import { useGlobalConfig } from '../../services/global-config/global-config'
+import { useGlobalConfig } from '../../services/global-config/use-global-confignfig'
 import { useColors } from '../../composables'
 import { VaButton } from '../va-button'
 import { VaRating } from '../va-rating/'
