--- packages/ui/src/components/va-config/WithConfigTransportTest.ts
@@ -1,4 +1,0 @@
-import ConfigUsageTest from './ConfigUsageTest.vue'
-import withConfigTransport from '../../services/config-transport/withConfigTransport'
-
-export default withConfigTransport(ConfigUsageTest as any)
