+++ packages/ui/src/components/va-sidebar/VaSidebar.vue
@@ -17,7 +17,7 @@
 <script lang="ts">
 import { defineComponent, computed, ref, PropType } from 'vue'
 
-import { getGradientBackground } from '../../services/color-config/color-functions'
+import { getGradientBackground } from '../../services/color-config'
 import { useColors, useTextColor, useBem } from '../../composables'
 import { useSidebar } from './hooks/useSidebar'
 import { useComponentPresetProp } from '../../composables/useComponentPreset'
