+++ packages/ui/src/composables/index.ts
@@ -61,3 +61,5 @@ export * from './useWindow'
 export * from './useWindowSize'
 export * from './useSlotPassed'
 export * from './useReactiveComputed'
+export * from './useIcon'
+export * from './useGlobalConfig'
