+++ packages/ui/src/composables/useGlobalConfig.ts
@@ -1,0 +1,1 @@
+export { useGlobalConfig } from '../services/global-config/global-config'
