+++ packages/ui/src/main.ts
@@ -1,17 +1,17 @@
 export * from './vuestic-plugin'
 export { useColors, useBreakpoint, useI18nConfig } from './composables'
-export { useGlobalConfig } from './services/global-config/global-config'
+export { useGlobalConfig } from './services/global-config'
 export {
-  useIcons,
+  useIcon as useIcons,
   createIconsConfig,
   VuesticIconFonts,
   VuesticIconAliases,
-} from './services/icon-config/icon-config'
+} from './services/icon-config'
 
 // Export all components and their composables
 export * from './components'
 
 export type { GlobalConfig, GlobalConfigUpdater, PartialGlobalConfig } from './services/global-config/types'
-export type { ComponentConfig } from './services/component-config/component-config'
+export type { ComponentConfig } from './services/component-config'
 export type { IconConfig, IconConfiguration } from './services/icon-config/types'
 export type { ColorConfig } from './services/color-config/types'
