+++ packages/ui/src/services/color-config/config/default.ts
@@ -1,0 +1,12 @@
+import type { ColorConfig } from '../types'
+import { presets } from '../presets'
+
+export const getColorDefaultConfig = (): ColorConfig => ({
+  variables: presets.light,
+  threshold: 150,
+  presets: {
+    light: presets.light,
+    dark: presets.dark,
+  },
+  currentPresetName: 'light',
+})
