+++ packages/ui/src/services/color-config/tests/color-functions.spec.ts
@@ -1,4 +1,4 @@
-import { setHSLAColor } from '../color-functions'
+import { setHSLAColor } from '../utils'
 import { describe, it, expect } from 'vitest'
 
 describe('ColorFunctions', () => {
