+++ packages/ui/src/services/component-config/index.ts
@@ -1,3 +1,3 @@
 export * from './types'
 export * from './config/default'
-export * from './utils'
+export * from './utils/use-component-config-props'
