+++ packages/ui/src/services/config-transport/index.ts
@@ -1,0 +1,1 @@
+export * from './withConfigTransport'
