+++ packages/ui/src/services/global-config/index.ts
@@ -1,0 +1,2 @@
+export * from './global-config'
+export * from './types'
