+++ packages/ui/src/services/i18n/index.ts
@@ -1,2 +1,2 @@
-export * from './defaults'
-export * from '../../composables/useI18n'
+export * from './config/default'
+export * from './types'
