+++ packages/ui/src/services/i18n/types.ts
@@ -1,0 +1,3 @@
+import type { getI18nConfigDefaults } from './config/default'
+
+export type I18nConfig = ReturnType<typeof getI18nConfigDefaults>
