+++ packages/ui/src/services/icon-config/config/default.ts
@@ -1,0 +1,4 @@
+import type { IconConfig } from '../types'
+import { createIconsConfig } from '../create-icons-config'
+
+export const getIconDefaultConfig = (): IconConfig => createIconsConfig({})
