+++ packages/ui/src/services/icon-config/create-icons-config.ts
@@ -1,6 +1,8 @@
 import { IconConfig, IconConfiguration } from './types'
 import { VuesticIconAliases, VuesticIconFonts } from './presets'
 
+// Helper for user to create icon config
+
 /**
  * Helper allow you to create icons config with Vuestic defaults.
  * Alias should have `to` property which is the name of font.
