--- packages/ui/src/services/icon-config/icon-config.ts
@@ -1,4 +1,0 @@
-export { useIcons } from './icon-config-hooks'
-export { createIconsConfig } from './icon-config-helpers'
-export { VuesticIconAliases, VuesticIconFonts } from './presets'
-export type { IconConfig } from './types'
