+++ packages/ui/src/services/icon-config/tests/utils/get-icon.spec.ts
@@ -1,5 +1,5 @@
-import { getIconConfiguration } from '../../icon-helpers'
-import { IconConfig } from '../../icon-config'
+import type { IconConfig } from '../../types'
+import { getIconConfiguration } from '../../utils/get-icon-configuration'
 import { describe, it, expect } from 'vitest'
 
 describe('IconConfig', () => {
